refactor: 당첨 번호 입력 검증 함수 중 공통되는 조건 하나의 함수만 사용하도록 리팩토링
