+++ src/main/java/lotto/validation/BonusNumberValidator.java
@@ -1,8 +1,6 @@
 package lotto.validation;
 
-import static lotto.util.LottoConstants.MAX_LOTTO_NUMBER;
-import static lotto.util.LottoConstants.MIN_LOTTO_NUMBER;
-import static lotto.validation.LottoNumberValidator.validateBonusNumber;
+import static lotto.validation.CommonLottoNumberValidator.validateLottoInput;
 
 import java.util.List;
 import lotto.domain.Lotto;
@@ -12,15 +10,7 @@ import lotto.view.input.InvalidInputException;
 public class BonusNumberValidator {
 
     public static void validate(String input) {
-        validateBonusNumber(input);
-        validateLottoRange(input);
-    }
-
-    private static void validateLottoRange(String input) {
-        int inputNumber = Integer.parseInt(input);
-        if (inputNumber < MIN_LOTTO_NUMBER.getValue() || inputNumber > MAX_LOTTO_NUMBER.getValue()) {
-            throw new InvalidInputException(InputErrorMessage.LOTTO_NUMBER_RANGE_INVALID);
-        }
+        validateLottoInput(input);
     }
 
     public static void validateDuplicate(Integer input, Lotto winningLotto) {
