+++ src/main/java/lotto/validation/LottoNumbersValidator.java
@@ -1,21 +1,31 @@
 package lotto.validation;
 
 import static lotto.util.LottoConstants.LOTTO_LENGTH;
-import static lotto.util.LottoConstants.MAX_LOTTO_NUMBER;
-import static lotto.util.LottoConstants.MIN_LOTTO_NUMBER;
+import static lotto.util.LottoConstants.ZERO_THRESHOLD;
+import static lotto.validation.CommonLottoNumberValidator.validateLottoInput;
 
+import java.util.Arrays;
 import java.util.List;
 import lotto.view.input.InputErrorMessage;
 import lotto.view.input.InvalidInputException;
 
 public class LottoNumbersValidator {
+    private final static String COMMA = ",";
 
-    public static void validate(List<Integer> numbers) {
-        validateLottoLength(numbers);
-        validateLottoRange(numbers);
-        validateDuplicate(numbers);
+    public static void validateAboutComma(String input) {
+        validateHasComma(input);
+        validateCommaSeparator(input);
+
     }
 
+    public static void validateWinningNumbers(List<Integer> input) {
+        validateLottoLength(input);
+        validateDuplicate(input);
+
+        for (Integer number : input) {
+            validateLottoInput(String.valueOf(number));
+        }
+    }
 
     private static void validateLottoLength(List<Integer> numbers) {
         if (numbers.size() != LOTTO_LENGTH.getValue()) {
@@ -23,15 +33,18 @@ public class LottoNumbersValidator {
         }
     }
 
-    private static void validateLottoRange(List<Integer> numbers) {
-        if (!isLottoRange(numbers)) {
-            throw new InvalidInputException(InputErrorMessage.LOTTO_NUMBER_RANGE_INVALID);
+    private static void validateHasComma(String input) {
+        if (!input.contains(COMMA)) {
+            throw new InvalidInputException(InputErrorMessage.LOTTO_NUMBER_SEPARATOR_MUST_BE_COMMA);
         }
     }
 
-    private static boolean isLottoRange(List<Integer> numbers) {
-        return numbers.stream()
-                .allMatch(number -> number >= MIN_LOTTO_NUMBER.getValue() && number <= MAX_LOTTO_NUMBER.getValue());
+    private static void validateCommaSeparator(String input) {
+        String[] inputs = input.split(COMMA);
+        long count = Arrays.stream(inputs).filter(String::isEmpty).count();
+        if (count > ZERO_THRESHOLD.getValue()) {
+            throw new InvalidInputException(InputErrorMessage.LOTTO_NUMBER_SEPARATOR_MUST_BE_COMMA);
+        }
     }
 
     private static void validateDuplicate(List<Integer> numbers) {
